Install signed images if UEFI Secure Boot is enabled
If UEFI is used and a signed shim (the first-stage boot loader) is available
as /boot/efi/EFI/rosa/BOOT*.EFI, use the latter as the default loader.
Otherwise, use the appropriate variant of GRUB: the common grub*.efi or the
version for removable devices (grubcd*.efi).

Also, generate a minimal grub.cfg in /boot/efi/EFI/rosa/ that, in turn,
pulls in the distro-specific configuration. This allows updating
/boot/grub2/grub.cfg as usual without having to write to /boot/efi/EFI/rosa/.

The patches used in ROSA with GRUB 2.00 for that purpose
(grub2-2.00-efi-install-secureboot-support.patch and
grub2-2.00-efi-install-grub-cfg.patch) cannot be reused with GRUB 2.02+
because grub-install was rewritten in C there.

This patch is based on the following one from Ubuntu, but it was reworked
to reflect ROSA-specific file paths, names and other conventions:

  From: Colin Watson <cjwatson@ubuntu.com>
  Date: Mon, 13 Jan 2014 12:13:22 +0000
  Subject: Install signed images if UEFI Secure Boot is enabled

  Author: Stéphane Graber <stgraber@ubuntu.com>
  Author: Steve Langasek <steve.langasek@ubuntu.com>
  Author: Linn Crosetto <linn@hpe.com>
  Author: Mathieu Trudel-Lapierre <cyphermox@ubuntu.com>

Signed-off-by: Evgenii Shatokhin <eshatokhin@virtuozzo.com>